Moved "currency_column_to_numeric" from functions to finance, shouldn't be needed outside of financial applications.
