+++ janitor/finance.py
@@ -4,7 +4,7 @@ Finance-specific data cleaning functions.
 
 import json
 from datetime import date, datetime
-from functools import lru_cache
+from functools import lru_cache, partial
 
 import pandas as pd
 import pandas_flavor as pf
@@ -12,7 +12,14 @@ import requests
 
 from janitor import check
 
-from .utils import deprecated_alias
+from .utils import (
+    deprecated_alias,
+    _clean_accounting_column,
+    _currency_column_to_numeric,
+    _replace_empty_string_with_none,
+    _replace_original_empty_string_with_none,
+)
+
 
 currency_set = {
     "AUD",
@@ -393,6 +400,233 @@ def _convert_currency(
     currency_dict = json.loads(result.text)
     rate = currency_dict["rates"][to_currency]
     return rate
+
+
+@pf.register_dataframe_method
+@deprecated_alias(col_name="column_name", type="cleaning_style")
+def currency_column_to_numeric(
+    df: pd.DataFrame,
+    column_name,
+    cleaning_style: str = None,
+    cast_non_numeric: dict = None,
+    fill_all_non_numeric: float = None,
+    remove_non_numeric: bool = False,
+) -> pd.DataFrame:
+    """
+    Convert currency column to numeric.
+
+    This method does not mutate the original DataFrame.
+
+    This method allows one to take a column containing currency values,
+    inadvertently imported as a string, and cast it as a float. This is
+    usually the case when reading CSV files that were modified in Excel.
+    Empty strings (i.e. `''`) are retained as `NaN` values.
+
+    :param df: The DataFrame
+    :param column_name: The column to modify
+    :param cleaning_style: What style of cleaning to perform. If None, standard
+        cleaning is applied. Options are:
+
+            * 'accounting':
+            Replaces numbers in parentheses with negatives, removes commas.
+
+    :param cast_non_numeric: A dict of how to coerce certain strings. For
+        example, if there are values of 'REORDER' in the DataFrame,
+        {'REORDER': 0} will cast all instances of 'REORDER' to 0.
+    :param fill_all_non_numeric: Similar to `cast_non_numeric`, but fills all
+        strings to the same value. For example,  fill_all_non_numeric=1, will
+        make everything that doesn't coerce to a currency 1.
+    :param remove_non_numeric: Will remove rows of a DataFrame that contain
+        non-numeric values in the `column_name` column. Defaults to `False`.
+    :returns: A pandas DataFrame.
+    """
+    # TODO: Convert this to a notebook.
+    # :Example Setup:
+
+    # .. code-block:: python
+
+    #     import pandas as pd
+    #     import janitor
+    #     data = {
+    #         "a": ["-$1.00", "", "REPAY"] * 2 + ["$23.00", "",
+    # "Other Account"],
+    #         "Bell__Chart": [1.234_523_45, 2.456_234, 3.234_612_5] * 3,
+    #         "decorated-elephant": [1, 2, 3] * 3,
+    #         "animals@#$%^": ["rabbit", "leopard", "lion"] * 3,
+    #         "cities": ["Cambridge", "Shanghai", "Basel"] * 3,
+    #     }
+    #     df = pd.DataFrame(data)
+
+    # :Example 1: Coerce numeric values in column to float:
+
+    # .. code-block:: python
+
+    #     df.currency_column_to_numeric("a")
+
+    # :Output:
+
+    # .. code-block:: python
+
+    #           a  Bell__Chart  decorated-elephant animals@#$%^     cities
+    #     0  -1.0     1.234523                   1       rabbit  Cambridge
+    #     1   NaN     2.456234                   2      leopard   Shanghai
+    #     2   NaN     3.234612                   3         lion      Basel
+    #     3  -1.0     1.234523                   1       rabbit  Cambridge
+    #     4   NaN     2.456234                   2      leopard   Shanghai
+    #     5   NaN     3.234612                   3         lion      Basel
+    #     6  23.0     1.234523                   1       rabbit  Cambridge
+    #     7   NaN     2.456234                   2      leopard   Shanghai
+    #     8   NaN     3.234612                   3         lion      Basel
+
+    # :Example 2: Coerce numeric values in column to float, and replace a
+    # string\
+    # value with a specific value:
+
+    # .. code-block:: python
+
+    #     cast_non_numeric = {"REPAY": 22}
+    #     df.currency_column_to_numeric("a", cast_non_numeric=cast_non_numeric)
+
+    # :Output:
+
+    # .. code-block:: python
+
+    #           a  Bell__Chart  decorated-elephant animals@#$%^     cities
+    #     0  -1.0     1.234523                   1       rabbit  Cambridge
+    #     1   NaN     2.456234                   2      leopard   Shanghai
+    #     2  22.0     3.234612                   3         lion      Basel
+    #     3  -1.0     1.234523                   1       rabbit  Cambridge
+    #     4   NaN     2.456234                   2      leopard   Shanghai
+    #     5  22.0     3.234612                   3         lion      Basel
+    #     6  23.0     1.234523                   1       rabbit  Cambridge
+    #     7   NaN     2.456234                   2      leopard   Shanghai
+    #     8   NaN     3.234612                   3         lion      Basel
+
+    # :Example 3: Coerce numeric values in column to float, and replace all\
+    #     string value with a specific value:
+
+    # .. code-block:: python
+
+    #     df.currency_column_to_numeric("a", fill_all_non_numeric=35)
+
+    # :Output:
+
+    # .. code-block:: python
+
+    #           a  Bell__Chart  decorated-elephant animals@#$%^     cities
+    #     0  -1.0     1.234523                   1       rabbit  Cambridge
+    #     1   NaN     2.456234                   2      leopard   Shanghai
+    #     2  35.0     3.234612                   3         lion      Basel
+    #     3  -1.0     1.234523                   1       rabbit  Cambridge
+    #     4   NaN     2.456234                   2      leopard   Shanghai
+    #     5  35.0     3.234612                   3         lion      Basel
+    #     6  23.0     1.234523                   1       rabbit  Cambridge
+    #     7   NaN     2.456234                   2      leopard   Shanghai
+    #     8  35.0     3.234612                   3         lion      Basel
+
+    # :Example 4: Coerce numeric values in column to float, replace a string\
+    #     value with a specific value, and replace remaining string values
+    # with\
+    #     a specific value:
+
+    # .. code-block:: python
+
+    #     df.currency_column_to_numeric("a", cast_non_numeric=cast_non_numeric,
+    #     fill_all_non_numeric=35)
+
+    # :Output:
+
+    # .. code-block:: python
+
+    #           a  Bell__Chart  decorated-elephant animals@#$%^     cities
+    #     0  -1.0     1.234523                   1       rabbit  Cambridge
+    #     1   NaN     2.456234                   2      leopard   Shanghai
+    #     2  22.0     3.234612                   3         lion      Basel
+    #     3  -1.0     1.234523                   1       rabbit  Cambridge
+    #     4   NaN     2.456234                   2      leopard   Shanghai
+    #     5  22.0     3.234612                   3         lion      Basel
+    #     6  23.0     1.234523                   1       rabbit  Cambridge
+    #     7   NaN     2.456234                   2      leopard   Shanghai
+    #     8  35.0     3.234612                   3         lion      Basel
+
+    # :Example 5: Coerce numeric values in column to float, and remove string\
+    #     values:
+
+    # .. code-block:: python
+
+    #     df.currency_column_to_numeric("a", remove_non_numeric=True)
+
+    # :Output:
+
+    # .. code-block:: python
+
+    #           a  Bell__Chart  decorated-elephant animals@#$%^     cities
+    #     0  -1.0     1.234523                   1       rabbit  Cambridge
+    #     1   NaN     2.456234                   2      leopard   Shanghai
+    #     3  -1.0     1.234523                   1       rabbit  Cambridge
+    #     4   NaN     2.456234                   2      leopard   Shanghai
+    #     6  23.0     1.234523                   1       rabbit  Cambridge
+    #     7   NaN     2.456234                   2      leopard   Shanghai
+
+    # :Example 6: Coerce numeric values in column to float, replace a string\
+    #     value with a specific value, and remove remaining string values:
+
+    # .. code-block:: python
+
+    #     df.currency_column_to_numeric("a", cast_non_numeric=cast_non_numeric,
+    #     remove_non_numeric=True)
+
+    # :Output:
+
+    # .. code-block:: python
+
+    #           a  Bell__Chart  decorated-elephant animals@#$%^     cities
+    #     0  -1.0     1.234523                   1       rabbit  Cambridge
+    #     1   NaN     2.456234                   2      leopard   Shanghai
+    #     2  22.0     3.234612                   3         lion      Basel
+    #     3  -1.0     1.234523                   1       rabbit  Cambridge
+    #     4   NaN     2.456234                   2      leopard   Shanghai
+    #     5  22.0     3.234612                   3         lion      Basel
+    #     6  23.0     1.234523                   1       rabbit  Cambridge
+    #     7   NaN     2.456234                   2      leopard   Shanghai
+
+    check("column_name", column_name, [str])
+
+    column_series = df[column_name]
+    if cleaning_style == "accounting":
+        df.loc[:, column_name] = df[column_name].apply(
+            _clean_accounting_column
+        )
+        return df
+
+    if cast_non_numeric:
+        check("cast_non_numeric", cast_non_numeric, [dict])
+
+    _make_cc_patrial = partial(
+        _currency_column_to_numeric, cast_non_numeric=cast_non_numeric
+    )
+
+    column_series = column_series.apply(_make_cc_patrial)
+
+    if remove_non_numeric:
+        df = df.loc[column_series != "", :]
+
+    # _replace_empty_string_with_none is applied here after the check on
+    # remove_non_numeric since "" is our indicator that a string was coerced
+    # in the original column
+    column_series = column_series.apply(_replace_empty_string_with_none)
+
+    if fill_all_non_numeric is not None:
+        check("fill_all_non_numeric", fill_all_non_numeric, [int, float])
+        column_series = column_series.fillna(fill_all_non_numeric)
+
+    column_series = column_series.apply(
+        _replace_original_empty_string_with_none
+    )
+
+    df = df.assign(**{column_name: pd.to_numeric(column_series)})
+
+    return df
 
 
 @pf.register_dataframe_method
